requirements.txt
altair==5.5.0
asn1crypto==1.5.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.38.44
botocore==1.38.44
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.3.0
cryptography==45.0.4
cycler==0.12.1
easyocr==1.7.2
entrypoints==0.4
et_xmlfile==2.0.0
Faker==37.4.0
favicon==0.7.0
filelock==3.16.0
fonttools==4.53.1
fsspec==2024.9.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
htbuilder==0.9.0
idna==3.10
imageio==2.35.1
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
lazy_loader==0.4
lxml==5.4.0
Markdown==3.8.2
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
mysql-connector-python==9.0.0
narwhals==1.44.0
networkx==3.3
ninja==1.11.1.1
numpy==2.1.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.3.0
pillow==10.4.0
platformdirs==4.3.8
plotly==6.1.2
prometheus_client==0.22.1
protobuf==5.29.5
pyarrow==18.1.0
pyclipper==1.3.0.post5
pycparser==2.22
pydeck==0.9.1
PyJWT==2.10.1
pymdown-extensions==10.16
pyOpenSSL==25.1.0
pyparsing==3.1.4
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
s3transfer==0.13.0
scikit-image==0.24.0
scipy==1.14.1
setuptools==74.1.2
shapely==2.0.6
six==1.16.0
smmap==5.0.2
snowflake-connector-python==3.15.0
snowflake-snowpark-python==1.33.0
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
st-annotated-text==4.0.2
st-theme==1.2.3
streamlit==1.46.0
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.7.1
streamlit-faker==0.0.3
streamlit-image-coordinates==0.3.1
streamlit-keyup==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
sympy==1.13.2
tenacity==9.1.2
tifffile==2024.8.30
toml==0.10.2
tomlkit==0.13.3
torch==2.4.1
torchvision==0.19.1
tornado==6.5.1
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
validators==0.35.0
watchdog==6.0.0
wheel==0.45.1
